% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kmeans}
\alias{kmeans}
\title{Returns a df of observations with thier assigned cluster}
\usage{
kmeans(x, y, clust, n)
}
\arguments{
\item{x}{the x variable to cluster on}

\item{y}{the y variable to cluster on}

\item{clust}{the number of clusters to create}

\item{n}{number of iterations for k-means to run}
}
\value{
df with assigned cluster
}
\description{
Returns a df of observations with thier assigned cluster
}
\examples{
data(iris)
kmeans(iris$Sepal.Length, iris$Sepal.Width, 4, 10)
}
