% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_shiny.R
\name{kmeans_shiny}
\alias{kmeans_shiny}
\title{Runs kmeans app on iris dataset
must have iris df loaded}
\usage{
kmeans_shiny(directory)
}
\arguments{
\item{directory}{the file directory that contains ui and server}
}
\value{
kmeans shiny app
}
\description{
Runs kmeans app on iris dataset
must have iris df loaded
}
\examples{
data(iris)
shiny_app("C:/Users/James/OneDrive/School/MATH_3190/github/hw2kmeans/inst")
}
